refactor(web): move timeline logic next to MessagesTimeline
