+++ apps/web/src/components/chat/MessagesTimeline.logic.ts
@@ -1,0 +1,25 @@
+export interface TimelineDurationMessage {
+  id: string;
+  role: "user" | "assistant" | "system";
+  createdAt: string;
+  completedAt?: string | undefined;
+}
+
+export function computeMessageDurationStart(
+  messages: ReadonlyArray<TimelineDurationMessage>,
+): Map<string, string> {
+  const result = new Map<string, string>();
+  let lastBoundary: string | null = null;
+
+  for (const message of messages) {
+    if (message.role === "user") {
+      lastBoundary = message.createdAt;
+    }
+    result.set(message.id, lastBoundary ?? message.createdAt);
+    if (message.role === "assistant" && message.completedAt) {
+      lastBoundary = message.completedAt;
+    }
+  }
+
+  return result;
+}
